% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_docs.R, R/stack_add.R, R/stack_rm.R
\name{add_rm}
\alias{add_rm}
\alias{stack_add}
\alias{stack_rm}
\title{Add or remove stack members}
\usage{
stack_add(stack, member, ...)

stack_rm(stack, member, ...)
}
\arguments{
\item{stack}{A model \code{stack} object.}

\item{member}{A \code{tune_results} object (or, if removing, its name)
outputted by \code{\link[tune:tune_grid]{tune::tune_grid()}} or \code{\link[tune:tune_bayes]{tune::tune_bayes()}}.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A model \code{stack} object--see \code{\link[=stack_init]{stack_init()}} for more details!
}
\description{
Add or remove models from a model stack specification.
}
\examples{
# See ?example_data to learn more about the
# `svm_res_` and `spline_res_` objects exported
# with the package. Long story short, they're example
# `tune_results` objects outputted from [tune::tune_grid()]!

# initialize a model stack
st <- stack_init()

# add some members to the stack
st <- st \%>\%
  stack_add(svm_res_) \%>\%
  stack_add(spline_res_)

# remove the spline regression stack member
st <- st \%>\%
  stack_rm("spline_res_")

}
