% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{augment.model_stack}
\alias{augment.model_stack}
\title{A barebones augment method to help with testing.}
\usage{
\method{augment}{model_stack}(x, data = x[["train"]], ...)
}
\arguments{
\item{x}{A \code{model_stack} object}

\item{data}{A \code{data.frame}-like object to collect predictions on.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
A barebones augment method to help with testing.
}
