% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.stack}
\alias{autoplot.stack}
\title{Plot an evaluated model stack}
\usage{
\method{autoplot}{stack}(
  object,
  type = c("marginals", "parameters", "performance"),
  metric = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{object}{An evaluated model stack.}

\item{type}{A single character value. Choices are \code{"marginals"} (for a plot
of each predictor versus performance; see Details below), \code{"parameters"}
(each parameter versus search iteration), or \code{"performance"} (performance
versus iteration). The latter two choices are only used for \code{\link[tune:tune_bayes]{tune_bayes()}}.}

\item{metric}{A character vector or \code{NULL} for which metric to plot. By
default, all metrics will be shown via facets.}

\item{width}{A number for the width of the confidence interval bars when
\code{type = "performance"}. A value of zero prevents them from being shown.}

\item{...}{For plots with a regular grid, this is passed to \code{format()} and is
applied to a parameter used to color points. Otherwise, it is not used.}
}
\description{
Plot an evaluated model stack
}
