% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_predictions.R
\name{blend_predictions}
\alias{blend_predictions}
\title{Evaluate a data stack}
\usage{
blend_predictions(data_stack, penalty = 10^(-6:-1), verbose = FALSE, ...)
}
\arguments{
\item{data_stack}{A \code{data_stack} object}

\item{penalty}{A numeric vector of proposed penalty values used in member
weighting. Higher penalties will generally result in fewer members
being included in the resulting model stack, and vice versa. This argument
will be tuned on unless a single penalty value is given.}

\item{verbose}{A logical for logging results as they are generated. Despite
this argument, warnings and errors are always shown.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{model_stack} object—while \code{model_stacks} largely contain the
same elements as \code{data_stack}s, the primary data objects shift from the
assessment set predictions to the member models.
}
\description{
Evaluates a stack by performing regularization on the out-of-sample
predictions to determine coefficients for the combining of predictions
from ensemble members.
}
\section{Example Data}{


These examples make use of data and resampling objects exported with the
package. All of them are derived from test cases making use of the
\code{penguins} dataset from Allison Horst's \code{palmerpenguins} package.

The \code{penguins_train} and \code{penguins_test} objects are subsets of the
penguins data for using in training and testing, respectively.

Objects containing the substring \verb{_res_} are \code{tune_results} objects
for model specifications. The \code{reg} prefix indicates that the model
definition is for use in regression, \code{class} indicates multinomial
classification, and \code{log} (as in logistic) indicates binomial classification.
The suffix indicates the model definition type. All of these \code{tune_results}
objects are fitted on a 5-fold cross validation of the \code{penguins_train} data.

For the regression setting, these \code{tune_results} objects reflect
models specified to fit \code{body_mass_g} using all of the other variables
as predictors. These objects include:
\itemize{
\item \code{reg_res_lr}: Fitted resamples for a linear regression model
\item \code{reg_res_sp}: Tuning results for a splines model
\item \code{reg_res_svm}: Tuning results for a support vector machine model
}

In the multinomial classification setting, the relevant objects reflect
models specified to fit \code{year} (as a factor) using all of the other variables
as predictors. These objects include:
\itemize{
\item \code{class_res_nn}: Fitted resamples for a neural network model
\item \code{class_res_rf}: Tuning results for a random forest model
}

In the binomial classification setting, the relevant objects reflect models
specified to fit \code{sex} using all of the other variables as predictors.
These objects include:
\itemize{
\item \code{log_res_nn}: Fitted resamples for a neural network model
\item \code{log_res_rf}: Tuning results for a random forest model
}

See \code{?example_data} to learn more about these objects, as well as browse
the source code that generated them.
}

\examples{
\donttest{
# see the "Example Data" section above for
# clarification on the objects used in these examples!

# put together a data stack
reg_st <- 
  stacks() \%>\%
  add_candidates(reg_res_lr) \%>\%
  add_candidates(reg_res_svm) \%>\%
  add_candidates(reg_res_sp)
  
reg_st

# evaluate the data stack
reg_st \%>\%
  blend_predictions()

# include fewer models by proposing
# higher penalties
reg_st \%>\% blend_predictions(penalty = c(.5, 1))
  
# do the same with multinomial classification models
class_st <-
  stacks() \%>\%
  add_candidates(class_res_nn) \%>\%
  add_candidates(class_res_rf) \%>\%
  blend_predictions()
  
class_st

# ...or binomial classification models
log_st <-
  stacks() \%>\%
  add_candidates(log_res_nn) \%>\%
  add_candidates(log_res_rf) \%>\%
  blend_predictions()
  
log_st
}

}
\seealso{
Other core verbs: 
\code{\link{add_candidates}()},
\code{\link{fit_members}()},
\code{\link{stacks}()}
}
\concept{core verbs}
