% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{build_linear_predictor}
\alias{build_linear_predictor}
\alias{build_linear_predictor._elnet}
\alias{build_linear_predictor._lognet}
\alias{build_linear_predictor._multnet}
\title{Creates an R expression for a linear predictor from a data frame of terms and
coefficients}
\usage{
build_linear_predictor(x, ...)

\method{build_linear_predictor}{`_elnet`}(x, ...)

\method{build_linear_predictor}{`_lognet`}(x, ...)

\method{build_linear_predictor}{`_multnet`}(x, ...)
}
\arguments{
\item{x}{An object that uses a \code{\link[glmnet:glmnet]{glmnet::glmnet()}} model and all numeric predictors.}

\item{...}{Not currently used.}
}
\value{
An R expression or a list of R expressions, depending on the type of
model being used.
}
\description{
Creates an R expression for a linear predictor from a data frame of terms and
coefficients
}
\keyword{internal}
