% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_stack_grid}
\alias{control_stack_grid}
\alias{control_stack_resamples}
\alias{control_stack_bayes}
\title{Control wrappers}
\usage{
control_stack_grid()

control_stack_resamples()

control_stack_bayes()
}
\value{
A \link[tune:control_grid]{tune::control_grid}, \link[tune:control_bayes]{tune::control_bayes},
or \link[tune:control_grid]{tune::control_resamples} object.
}
\description{
Supply these light wrappers as the \code{control} argument in a
\link[tune:tune_grid]{tune::tune_grid}, \link[tune:tune_bayes]{tune::tune_bayes}, or \link[tune:fit_resamples]{tune::fit_resamples}
call to return the needed workflow elements for use in a data stack.
These functions will return the appropriate control grid to ensure that
a minimal workflow giving details on model specifications and
preprocessors is supplied in the resampling results object!
}
\seealso{
See \link{example_data} for examples of these functions used in context.
}
