% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_docs.R
\docType{data}
\name{example_data}
\alias{example_data}
\alias{svm_res_}
\alias{spline_res_}
\title{Example Data}
\format{
An object of class \code{tune_results} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 5 columns.

An object of class \code{tune_results} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 5 columns.
}
\usage{
svm_res_

spline_res_
}
\description{
This package provides some \code{tune_results} objects for use in examples
and tests. \link{svm_res_} and \link{spline_res_} contain tuning results
for a support vector machine and spline model, respectively, fitting
\code{mpg} in the \code{mtcars} data. The source code for generating
these objects is given below.
}
\examples{
\dontrun{
# setup: resample and basic recipe ------------------------

set.seed(1)

folds_ <- rsample::vfold_cv(mtcars, v = 3)

car_rec_ <- 
  recipes::recipe(mpg ~ ., data = mtcars) \%>\%
  recipes::step_normalize(recipes::all_predictors())

# support vector machine ----------------------------------

svm_mod_ <- 
  parsnip::svm_rbf(
    cost = tune::tune(), 
    rbf_sigma = tune::tune()
  ) \%>\%
  parsnip::set_engine("kernlab") \%>\%
  parsnip::set_mode("regression")

set.seed(1)

svm_res_ <- 
  tune::tune_grid(
    object = svm_mod_, 
    preprocessor = car_rec_, 
    resamples = folds_, 
    grid = 5
  )

# spline regression ---------------------------------------

spline_rec_ <-
  recipes::recipe(mpg ~ ., data = mtcars) \%>\%
  recipes::step_ns(disp, deg_free = tune::tune("disp")) \%>\%
  recipes::step_ns(wt, deg_free = tune::tune("wt"))

lin_mod_ <-
  parsnip::linear_reg() \%>\%
  parsnip::set_engine("lm")

spline_grid_ <- expand.grid(disp = c(2, 4, 6), wt = c(2, 4, 6))

spline_res_ <- 
  tune::tune_grid(
    object = lin_mod_,
    preprocessor = spline_rec_, 
    resamples = folds_, 
    grid = spline_grid_
  )
}
}
\keyword{datasets}
