% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.stacked_data}
\alias{fit.stacked_data}
\title{Fit an evaluated model stack.}
\usage{
\method{fit}{stacked_data}(object, ...)
}
\arguments{
\item{object}{Stacked data.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
Fit an evaluated model stack.
}
