% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{get_expressions}
\alias{get_expressions}
\alias{get_expressions._multnet}
\alias{get_expressions._lognet}
\alias{get_expressions._elnet}
\title{Obtain prediction equations for all possible values of type}
\usage{
get_expressions(x, ...)

\method{get_expressions}{`_multnet`}(x, ...)

\method{get_expressions}{`_lognet`}(x, ...)

\method{get_expressions}{`_elnet`}(x, ...)
}
\arguments{
\item{x}{A \code{parsnip} model with the \code{glmnet} engine.}

\item{...}{Not used}
}
\value{
A named list with prediction equations for each possibel type.
}
\description{
Obtain prediction equations for all possible values of type
}
