% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stack.R
\name{model_stack}
\alias{model_stack}
\title{Initialize a model stack}
\usage{
model_stack(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A model \code{stack} object.
}
\description{
This function initializes a model \code{stack} object. Principally, \code{stack}
objects contain out-of-sample predictions from ensemble members as well
as metadata identifying their connections.
}
\seealso{
\code{\link[=members_add]{members_add()}}, \code{\link[=members_rm]{members_rm()}}
}
