% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_stack.R
\name{new_stack}
\alias{new_stack}
\title{Initialize a stack}
\usage{
new_stack(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{stack} object.
}
\description{
This function initializes a \code{stack} object. Principally, \code{stack}
objects contain out-of-sample predictions from ensemble members as well
as metadata identifying their connections.
}
\seealso{
\code{\link[=members_add]{members_add()}}, \code{\link[=members_rm]{members_rm()}}
}
