% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancakes.R
\name{pancakes}
\alias{pancakes}
\title{Initialize a stack}
\usage{
pancakes(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{stack} object.
}
\description{
This function initializes a \code{stack} object. Principally, \code{stack}
objects contain out-of-sample predictions from ensemble members as well
as metadata identifying their connections.
}
\seealso{
\code{\link[=stack_resamples]{stack_resamples()}}, \code{\link[=remove_members]{remove_members()}}
}
