% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.model_stack}
\alias{predict.model_stack}
\title{Model predictions}
\usage{
\method{predict}{model_stack}(object, new_data, type = NULL, opts = list(), ...)
}
\arguments{
\item{object}{A model stack with fitted members.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{Format of returned predicted values—passed on to
\link[parsnip:predict.model_fit]{parsnip::predict.model_fit} for each member.}

\item{opts}{A list of optional arguments to the underlying predict
function passed on to \link[parsnip:predict.model_fit]{parsnip::predict.model_fit} for each member.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
Apply a model stack to create different types of predictions.
}
