% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{prediction_eqn}
\alias{prediction_eqn}
\alias{prediction_eqn._lognet}
\alias{prediction_eqn._elnet}
\alias{prediction_eqn._multnet}
\title{Convert one or more linear predictor to a format used for prediction}
\usage{
prediction_eqn(x, ...)

\method{prediction_eqn}{`_lognet`}(x, type = "class", ...)

\method{prediction_eqn}{`_elnet`}(x, type = "numeric", ...)

\method{prediction_eqn}{`_multnet`}(x, type = "class", ...)
}
\arguments{
\item{x}{An object that uses a \code{\link[glmnet:glmnet]{glmnet::glmnet()}} model and all numeric predictors.}

\item{...}{Not currently used.}

\item{type}{The prediction type.}
}
\value{
The return type varies, based on the model and prediction type.
}
\description{
Convert one or more linear predictor to a format used for prediction
}
\keyword{internal}
