% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{fit}}}
}}

