% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_coefficients.R
\name{stack_coefficients}
\alias{stack_coefficients}
\title{Evaluate a data stack}
\usage{
stack_coefficients(stack_data, ...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}

\item{data_stack}{A \code{data_stack} object outputted from \code{\link[=stack_data]{stack_data()}}}
}
\value{
A \code{data_stack} object---a \code{model_fit} subclass, giving loading
coefficients for each ensemble member.
}
\description{
Evaluates a data stack by performing regularization on the out-of-sample
predictions to determine coefficients for the combining of predictions
from ensemble members.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# evaluate a resample stack
st <- 
  stacks() \%>\%
  add_members(lin_reg_res_) \%>\%
  add_members(svm_res_) \%>\%
  add_members(spline_res_) \%>\% 
  stack_data()
  
# evaluate the data stack
st \%>\%
  stack_coefficients()

}
