% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\title{Evaluate a resample stack}
\usage{
stack_data(resample_stack)
}
\value{
A \code{data_stack} object---a \code{tbl_df} subclass, where the first column
gives the true outcome in the training data, and the remaining columns give
the out-of-sample predictions from each potential ensemble member.
}
\description{
Evaluates a resample stack by collating out-of-sample predictions
from each potential ensemble member.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a resample stack and add some members
st <- 
  stacks() \%>\%
  stack_resamples(lin_reg_res_) \%>\%
  stack_resamples(svm_res_) \%>\%
  stack_resamples(spline_res_)
  
# evaluate the resample stack
st \%>\% 
  stack_data()
  
}
