% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ensemble.R
\name{stack_ensemble}
\alias{stack_ensemble}
\title{Fit a stack ensemble}
\usage{
stack_ensemble(stack, data)
}
\arguments{
\item{stack}{A \code{stack} object.}
}
\value{
A \code{ensemble_stack} object---this fitted model contains the necessary
components to predict on new data.
}
\description{
Fit a stack ensemble
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a stack
st_res <- 
  stacks() \%>\%
  add_members(lin_reg_res_, lin_reg_wf_) \%>\%
  add_members(svm_res_, svm_wf_) \%>\%
  add_members(spline_res_, spline_wf_) \%>\%
  stack_coefficients()

# evaluate and build an ensemble
stack_ensemble(st_res, st_coefs)

}
