% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_eval.R
\name{stack_eval}
\alias{stack_eval}
\title{Evaluate model stack coefficients}
\usage{
stack_eval(stack, preprocessor, ...)
}
\arguments{
\item{stack}{A model \code{stack} object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{

}
\description{
Completes a stacked model specification by estimating the
stacking/loading coefficients for the model stack.
}
\examples{

#example man-roxygen/ex-note_example_data

# initialize a model stack and add some members
st <- stack_init() \%>\%
  stack_add(svm_res_) \%>\%
  stack_add(spline_res_)

# evaluate the stack
st \%>\%
  stack_eval()

}
