% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fit.R
\name{stack_fit}
\alias{stack_fit}
\title{Fit model stack members}
\usage{
stack_fit(model_stack, data = NULL, ...)
}
\arguments{
\item{model_stack}{A \code{model_stack} object outputted by \code{stack_blend()} or
\code{stack_fit()}}

\item{data}{The data used in generating the original tuning/fitting results.
Can generally be extracted from the \code{model_stack} object, so defaults to
\code{NULL}.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{model_stack} object with a subclass inherited from the chosen
\verb{*_stack} method---this fitted model contains the
necessary components to predict on new data.
}
\description{
Fit model stack members
}
\section{Example Data}{


These examples make use of data and resampling objects exported with the
package. All of them are derived from test cases making use of the
\code{penguins} dataset from Allison Horst's \code{palmerpenguins} package.

The \code{penguins_train} and \code{penguins_test} objects are subsets of the
penguins data for using in training and testing, respectively.

Objects containing the substring \verb{_res_} are \code{tune_results} objects
for model specifications. The \code{reg} prefix indicates that the model
definition is for use in regression, \code{class} indicates multinomial
classification, and \code{log} (as in logistic) indicates binomial classification.
The suffix indicates the model definition type. All of these \code{tune_results}
objects are fitted on a 5-fold cross validation of the \code{penguins_train} data.

For the regression setting, these \code{tune_results} objects reflect
models specified to fit \code{body_mass_g} using all of the other variables
as predictors. These objects include:
\itemize{
\item \code{reg_res_lr}: Fitted resamples for a linear regression model
\item \code{reg_res_sp}: Tuning results for a splines model
\item \code{reg_res_svm}: Tuning results for a support vector machine model
}

In the multinomial classification setting, the relevant objects reflect
models specified to fit \code{island} using all of the other variables as
predictors. These objects include:
\itemize{
\item \code{class_res_nn}: Fitted resamples for a neural network model
\item \code{class_res_rf}: Tuning results for a random forest model
}

In the binomial classification setting, the relevant objects reflect models
specified to fit \code{sex} using all of the other variables as predictors.
These objects include:
\itemize{
\item \code{log_res_nn}: Fitted resamples for a neural network model
\item \code{log_res_rf}: Tuning results for a random forest model
}

See \code{?example_data} to learn more about these objects, as well as browse
the source code that generated them.
}

\examples{
\donttest{
# put together a data stack
reg_st <- 
  stacks() \%>\%
  stack_add(reg_res_lr) \%>\%
  stack_add(reg_res_svm) \%>\%
  stack_add(reg_res_sp)

# evaluate the data stack and fit the member models
reg_st \%>\%
  stack_blend() \%>\%
  stack_fit()
  
# do the same with multinomial classification models
class_st <-
  stacks() \%>\%
  stack_add(class_res_nn) \%>\%
  stack_add(class_res_rf) \%>\%
  stack_blend() \%>\%
  stack_fit()
  
# ...or binomial classification models
log_st <-
  stacks() \%>\%
  stack_add(log_res_nn) \%>\%
  stack_add(log_res_rf) \%>\%
  stack_blend() \%>\%
  stack_fit()
}

}
\seealso{
Other core verbs: 
\code{\link{stack_add}()},
\code{\link{stack_blend}()},
\code{\link{stacks}()}
}
\concept{core verbs}
