% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_init.R
\name{stack_init}
\alias{stack_init}
\title{Initialize a model stack}
\usage{
stack_init(...)
}
\arguments{
\item{...}{Currently ignored.}
}
\value{
A model \code{stack} object.
}
\description{
This function initializes a model \code{stack} object. Principally, \code{stack}
objects contain a hash for the resampling object and a list of member
models.
}
\seealso{
\code{\link[=stack_add]{stack_add()}}, \code{\link[=stack_rm]{stack_rm()}}
}
