% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_linear.R
\name{stack_linear}
\alias{stack_linear}
\title{Evaluate a data stack}
\usage{
stack_linear(data_stack, ...)
}
\arguments{
\item{data_stack}{A \code{data_stack} object}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{model_stack} object—while \code{model_stacks} largely contain the
same elements as \code{data_stack}s, the primary data objects shift from the
assessment set predictions to the member models.
}
\description{
Evaluates a stack by performing regularization on the out-of-sample
predictions to determine coefficients for the combining of predictions
from ensemble members.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{reg_res_svm}, and \code{reg_res_sp} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# put together a data stack
st <- 
  stacks() \%>\%
  add_candidates(reg_res_lr) \%>\%
  add_candidates(reg_res_svm) \%>\%
  add_candidates(reg_res_sp)
  
# evaluate the data stack
st \%>\%
  stack_linear()

}
