% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{stack_predict}
\alias{stack_predict}
\alias{stack_predict.elnet_numeric}
\alias{stack_predict.lognet_class}
\alias{stack_predict.lognet_prob}
\alias{stack_predict.multnet_class}
\alias{stack_predict.multnet_prob}
\title{Convert one or more linear predictor to a format used for prediction}
\usage{
stack_predict(x, ...)

\method{stack_predict}{elnet_numeric}(x, data, ...)

\method{stack_predict}{lognet_class}(x, data, ...)

\method{stack_predict}{lognet_prob}(x, data, ...)

\method{stack_predict}{multnet_class}(x, data, ...)

\method{stack_predict}{multnet_prob}(x, data, ...)
}
\arguments{
\item{x}{A set of model expressions generated by \code{\link[=prediction_eqn]{prediction_eqn()}}.}

\item{...}{Not currently used.}
}
\value{
The return type varies, based on the model and prediction type.
}
\description{
Convert one or more linear predictor to a format used for prediction
}
\keyword{internal}
