% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_predictions.R
\name{stack_predictions}
\alias{stack_predictions}
\title{Collate predictions from a model stack}
\usage{
stack_predictions(stack, ...)
}
\arguments{
\item{stack}{A model \code{stack} object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A tibble with \code{nrow(data)} rows and \verb{1 + <number of sub-models>}
columns, where each column (besides the first, which contains the true
response) gives the prediction for each given sub-model.
}
\description{
Given a model stack, this function collects the predictions
from each of the members' sub-models and collates them in a tibble.
This is a helper function and will generally not be used (explicitly)
in a model stacking pipeline, but is exported for user convenience
for ad-hoc model stacking.
}
