% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_resamples.R
\name{stack_resamples}
\alias{stack_resamples}
\title{Add members to a resample stack}
\usage{
stack_resamples(stack, members, name = deparse(substitute(members)), ...)
}
\arguments{
\item{stack}{A \code{resample_stack} object.}

\item{members}{A model definition: either a \code{tune_results}
or \code{resample_results} object outputted from
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}}

\item{name}{The label for the model definition---defaults to the name
of the model definition object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{resample_stack} object--see \code{\link[=stacks]{stacks()}} for more details!
}
\description{
Add the workflow and resampling objects for potential ensemble members
to a resample stack.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a model stack
st <- stacks()

# add some members to the stack
st <- st \%>\%
  stack_resample(lin_reg_res_) \%>\%
  stack_resamples(svm_res_) \%>\%
  stack_resamples(spline_res_)
  
}
