% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{stack_workflow}
\alias{stack_workflow}
\title{Compile a minimal workflow for use in stacks}
\usage{
stack_workflow(x)
}
\arguments{
\item{x}{A \code{workflow} object.}
}
\value{
A minimal \code{workflow} object with only the model specification and
preprocessor.
}
\description{
Supply this helper function as the \code{extract} argument in a
\link[tune:control_grid]{tune::control_grid}, \link[tune:control_bayes]{tune::control_bayes}, or \link[tune:control_grid]{tune::control_resamples}
call to return the needed workflow elements for use in a data stack.
}
\examples{
\dontrun{
library(tune)

tune::control_grid(save_pred = TRUE,
                   extract = stack_workflow)
}

}
\seealso{
See \link{control_stack_grid}, \link{control_stack_bayes},
and \link{control_stack_resamples} for wrappers around this function that
can be supplied directly to \link[tune:control_grid]{tune::control_grid}, \link[tune:control_bayes]{tune::control_bayes},
or \link[tune:control_grid]{tune::control_resamples}, respectively
}
