% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks}
\alias{stacks}
\title{Initialize a data stack}
\usage{
stacks(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{data_stack} object.
}
\description{
Initializes a \code{data_stack} object. Principally, \code{data_stack}s are just
tibbles, where the first column gives the true outcome in the assessment set,
and the remaining columns give the predictions from each candidate ensemble
member. (When the outcome is numeric, there’s only one column per candidate
member. For classification, there are as many columns per candidate
member as there are levels in the outcome variable.) They also bring
along a few extra attributes to keep track of model definitions, resamples,
and training data.
}
\seealso{
\code{\link[=stack_add]{stack_add()}}
}
