% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks_package.R
\docType{package}
\name{stacks_description}
\alias{stacks_description}
\alias{stacks_package}
\alias{stacks-package}
\title{stacks: Tidy Model Stacking}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Model stacking is an ensemble technique that involves training a model to combine the outputs of many diverse statistical models, and has been shown to improve predictive performance in a variety of settings. 'stacks' implements a grammar for 'tidymodels'-aligned model stacking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stacks.tidymodels.org/}
  \item \url{https://github.com/tidymodels/stacks}
  \item Report bugs at \url{https://github.com/tidymodels/stacks/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
