% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_candidates.R
\name{add_candidates}
\alias{add_candidates}
\title{Add model definitions to a data stack}
\usage{
add_candidates(
  data_stack,
  candidates,
  name = deparse(substitute(candidates)),
  ...
)
}
\arguments{
\item{data_stack}{A \code{data_stack} object.}

\item{candidates}{A model definition: either a \code{tune_results}
or \code{resample_results} object outputted from
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}.
These results must have been fitted with the \code{control} settings
\verb{save_pred = TRUE, save_workflow = TRUE}—see the \code{\link[=control_stack_grid]{control_stack_grid()}},
\code{\link[=control_stack_bayes]{control_stack_bayes()}}, and \code{\link[=control_stack_resamples]{control_stack_resamples()}}
documentation for helper functions.}

\item{name}{The label for the model definition---defaults to the name
of the \code{candidates} object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{data_stack} object--see \code{\link[=stacks]{stacks()}} for more details!
}
\description{
Collates assessment set predictions and appends workflows and additional
attributes to a \code{data_stack}.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{reg_res_svm}, and \code{reg_res_sp} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a stack and add some candidate members
st <- 
  stacks() \%>\%
  add_candidates(reg_res_lr) \%>\%
  add_candidates(reg_res_svm) \%>\%
  add_candidates(reg_res_sp)
  
# do the same with classification models
st2 <- 
  stacks() \%>\%
  add_candidates(class_res_nn) \%>\%
  add_candidates(class_res_rf)  
}
