% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_members.R
\name{add_members}
\alias{add_members}
\title{Add model definitions to a data stack}
\usage{
add_members(
  data_stack,
  members,
  workflow,
  name = deparse(substitute(members)),
  ...
)
}
\arguments{
\item{data_stack}{A \code{data_stack} object.}

\item{members}{A model definition: either a \code{tune_results}
or \code{resample_results} object outputted from
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}}

\item{workflow}{The workflow used to define the model definition.}

\item{name}{The label for the model definition---defaults to the name
of the \code{members} object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{data_stack} object--see \code{\link[=stacks]{stacks()}} for more details!
}
\description{
Collates assessment set predictions and appends workflows and additional
attributes to a \code{data_stack}.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{reg_res_svm}, and \code{reg_res_sp} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a stack and add some members
st <- 
  stacks() \%>\%
  add_members(reg_res_lr, reg_wf_lr) \%>\%
  add_members(reg_res_svm, reg_wf_svm) \%>\%
  add_members(reg_res_sp, reg_wf_sp)
  
# do the same with classification models
st <- 
  stacks() \%>\%
  add_members(class_res_nn, class_wf_nn) \%>\%
  add_members(class_res_rf, class_wf_rf)  
}
