% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_members.R
\name{add_members}
\alias{add_members}
\title{Add members to a resample stack}
\usage{
add_members(stack, members, workflow, name = deparse(substitute(members)), ...)
}
\arguments{
\item{stack}{A \code{data_stack} object.}

\item{members}{A model definition: either a \code{tune_results}
or \code{resample_results} object outputted from
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}}

\item{workflow}{The workflow used to define the model definition.}

\item{name}{The label for the model definition---defaults to the name
of the model definition object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{data_stack} object--see \code{\link[=stacks]{stacks()}} for more details!
}
\description{
Add the workflow and resampling objects for potential ensemble members
to a resample stack.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a stack and add some members
st <- 
  stacks() \%>\%
  add_members(lin_reg_res_, lin_reg_wf_) \%>\%
  add_members(svm_res_, svm_wf_) \%>\%
  add_members(spline_res_, spline_wf_)
  
}
