% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_members.R, R/misc_docs.R,
%   R/remove_members.R
\name{add_members}
\alias{add_members}
\alias{add_rm}
\alias{remove_members}
\title{Add or remove stack members}
\usage{
add_members(stack, members, name = deparse(substitute(members)), ...)

remove_members(stack, name, ...)
}
\arguments{
\item{stack}{A \code{stack} object.}

\item{...}{Additional arguments. Currently ignored.}

\item{member}{A \code{tune_results} object (or, if removing, its name)
outputted by \code{\link[tune:tune_grid]{tune::tune_grid()}} or \code{\link[tune:tune_bayes]{tune::tune_bayes()}}.}
}
\value{
A \code{stack} object--see \code{\link[=pancakes]{pancakes()}} for more details!
}
\description{
Add or remove members from a stack.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a model stack
st <- pancakes()

# add some members to the stack
st <- st \%>\%
  add_members(svm_res_) \%>\%
  add_members(spline_res_)

# remove the spline regression stack member
st <- st \%>\%
  remove_members("spline_res_")

}
