% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members_add.R, R/members_rm.R, R/misc_docs.R
\name{members_add}
\alias{members_add}
\alias{members_rm}
\alias{add_rm}
\title{Add or remove stack members}
\usage{
members_add(stack, members, name = deparse(substitute(members)), ...)

members_rm(stack, members, ...)
}
\arguments{
\item{stack}{A model \code{stack} object.}

\item{...}{Additional arguments. Currently ignored.}

\item{member}{A \code{tune_results} object (or, if removing, its name)
outputted by \code{\link[tune:tune_grid]{tune::tune_grid()}} or \code{\link[tune:tune_bayes]{tune::tune_bayes()}}.}
}
\value{
A model \code{stack} object--see \code{\link[=model_stack]{model_stack()}} for more details!
}
\description{
Add or remove models from a model stack specification.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{svm_res_} and
\code{spline_res_} objects used in the examples below. Long story short,
they're \code{tune_results} objects outputted from \code{\link[tune:tune_grid]{tune::tune_grid()}}
fitting \code{mpg} with all other variables in \code{mtcars} as predictors!
}

\examples{

# initialize a model stack
st <- model_stack()

# add some members to the stack
st <- st \%>\%
  members_add(svm_res_) \%>\%
  members_add(spline_res_)

# remove the spline regression stack member
st <- st \%>\%
  members_rm("spline_res_")

}
