% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{augment.model_stack}
\alias{augment.model_stack}
\title{Augment data with information from a \code{model_stack} object}
\usage{
\method{augment}{model_stack}(x, data = x[["train"]], ...)
}
\arguments{
\item{x}{A \code{model_stack} object}

\item{data}{A \code{data.frame}-like object containing all of the necessary
predictors that the model stack was trained with.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with additional information on each observation
appended.
}
\description{
Augment accepts a model object and a dataset and adds information about
each observation in the dataset.
}
\section{Example Data}{


This package provides some resampling objects and datasets for use in examples
and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal
7ish days if they detect potential predator threat. Researchers wanted
to determine how, and when, these tree frog embryos were able to detect
stimulus from their environment. To do so, they subjected the embryos
at varying developmental stages to "predator stimulus" by jiggling
the embryos with a blunt probe. Beforehand, though some of the embryos
were treated with gentamicin, a compound that knocks out their lateral
line (a sensory organ.) Researcher Julie Jung and her crew found that
these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily
a representative or unbiased subset of the complete dataset, and is
only for demonstrative purposes.

\code{reg_folds} and \code{class_folds} are \code{rset} cross-fold validation objects
from \code{rsample}, splitting the training data into for the regression
and classification model objects, respectively. \code{tree_frogs_reg_test} and
\code{tree_frogs_class_test} are the analogous testing sets.

\code{reg_res_lr}, \code{reg_res_svm}, and \code{reg_res_sp} contain regression tuning results
for a linear regression, support vector machine, and spline model, respectively,
fitting \code{latency} (i.e. how long the embryos took to hatch in response
to the jiggle) in the \code{tree_frogs} data, using most all of the other
variables as predictors. Note that the data underlying these models is
filtered to include data only from embryos that hatched in response to
the stimulus.

\code{class_res_rf} and \code{class_res_nn} contain multiclass classification tuning
results for a random forest and neural network classification model,
respectively, fitting \code{reflex} (a measure of ear function) in the
data using most all of the other variables as predictors.

\code{log_res_rf} and \code{log_res_nn}, contain binary classification tuning results
for a random forest and neural network classification model, respectively,
fitting \code{hatched} (whether or not the embryos hatched in response
to the stimulus) using most all of the other variables as predictors.

See \code{?example_data} to learn more about these objects, as well as browse
the source code that generated them.
}

\examples{
\donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

# to use the `tidy` method, the appropriate glance
# methods must be loaded for the member model tidiers.
# in this example, they come from broom!
library(broom)

# build a regression model stack
st <-
  stacks() \%>\%
  add_candidates(reg_res_lr) \%>\%
  add_candidates(reg_res_sp) \%>\%
  blend_predictions() \%>\%
  fit_members()
  
# check the model stack out!
tidy(st)
glance(st)
augment(st)
}

}
\seealso{
Other tidiers: 
\code{\link{glance.model_stack}()},
\code{\link{tidy.model_stack}()}
}
\concept{tidiers}
