% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{augment.model_stack}
\alias{augment.model_stack}
\title{Augment a model stack}
\usage{
\method{augment}{model_stack}(x, new_data, ...)
}
\arguments{
\item{x}{A fitted model stack; see \code{\link[=fit_members]{fit_members()}}.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{predict.model_stack}. In
particular, see \code{type} and \code{members}.}
}
\description{
Augment a model stack
}
\seealso{
The \code{\link[=collect_parameters]{collect_parameters()}} function is analogous to a \code{\link[generics:tidy]{tidy()}}
method for model stacks.
}
