% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.stack}
\alias{fit.stack}
\title{Fit an evaluated model stack.}
\usage{
\method{fit}{stack}(object, train_dat, ...)
}
\arguments{
\item{object}{Stacked data.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
Fit an evaluated model stack.
}
