% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_members.R
\name{fit_members}
\alias{fit_members}
\title{Fit model stack members}
\usage{
fit_members(model_stack, data = NULL, ...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{model_stack} object with a subclass inherited from the chosen
\verb{*_stack} method---this fitted model contains the
necessary components to predict on new data.
}
\description{
Fit model stack members
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a linear model stack
st_res <- 
  stacks() \%>\%
  add_members(lin_reg_res_, lin_reg_wf_) \%>\%
  add_members(svm_res_, svm_wf_) \%>\%
  add_members(spline_res_, spline_wf_) \%>\%
  linear_stack()

# fit the member models
fit_members(st_res, st_coefs)

}
