% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.data_stack}
\alias{predict.data_stack}
\title{Apply a stacked_ensemble to create different types of predictions.}
\usage{
\method{predict}{data_stack}(object, ...)
}
\arguments{
\item{object}{A data stack.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
Apply a stacked_ensemble to create different types of predictions.
}
