% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.model_stack}
\alias{predict.model_stack}
\title{Model predictions}
\usage{
\method{predict}{model_stack}(object, new_data, ...)
}
\arguments{
\item{object}{A model stack with fitted members.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments. Currently ignored.}
}
\description{
Apply a model stack to create different types of predictions.
}
