% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butcher.R, R/plots.R, R/predict.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{butcher}
\alias{axe_call}
\alias{axe_ctrl}
\alias{axe_data}
\alias{axe_env}
\alias{axe_fitted}
\alias{autoplot}
\alias{augment}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{butcher}{\code{\link[butcher]{axe_call}}, \code{\link[butcher]{axe_ctrl}}, \code{\link[butcher]{axe_data}}, \code{\link[butcher]{axe_env}}, \code{\link[butcher]{axe_fitted}}, \code{\link[butcher]{butcher}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{generics}{\code{\link[generics]{augment}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

