% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ensemble.R
\name{stack_ensemble}
\alias{stack_ensemble}
\title{Fit a stack ensemble}
\usage{
stack_ensemble(resample_stack, coefficient_stack)
}
\arguments{
\item{resample_stack}{A \code{resample_stack} object.}

\item{coefficient_stack}{A coefficient stack outputted by \code{\link[=stack_coefficients]{stack_coefficients()}}}
}
\value{
A \code{ensemble_stack} object---this fitted model contains the necessary
components to predict on new data.
}
\description{
Fit a stack ensemble
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{svm_res_}, and \code{spline_res_} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# initialize a resample stack and add some members
st_res <- 
  stacks() \%>\%
  stack_resamples(lin_reg_res_) \%>\%
  stack_resamples(svm_res_) \%>\%
  stack_resamples(spline_res_)
  
# evaluate the resample stack
st_coefs <-
  st_res \%>\% 
  stack_data() \%>\%
  stack_coefficients()
  
stack_ensemble(st_res, st_coefs)

}
