% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_eval.R
\name{stack_eval}
\alias{stack_eval}
\alias{stack_uneval}
\title{Evaluate model stack coefficients}
\usage{
stack_eval(stack, data, ...)

stack_uneval(stack, ...)
}
\arguments{
\item{stack}{A model \code{stack} object.}

\item{data}{The training data used to generate the resampling object.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
An evaluated model \code{stack} object.
}
\description{
Completes a stacked model specification by estimating the
stacking/loading coefficients for the model stack. Model stack
evaluation assumes that the model members will no longer be
altered—to adjust member models, remove the model evaluation
with \code{stack_uneval()}.
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{svm_res_} and
\code{spline_res_} objects used in the examples below. Long story short,
they're \code{tune_results} objects outputted from \code{\link[tune:tune_grid]{tune::tune_grid()}}
fitting \code{mpg} with all other variables in \code{mtcars} as predictors!
}

\examples{
# initialize a model stack and add some members
st <- stack_init() \%>\%
  stack_add(svm_res_) \%>\%
  stack_add(spline_res_)

# evaluate the stack
st \%>\%
  stack_eval(mtcars)
  
# the following code will result in warnings about
# immutability of evaluated model stacks
\dontrun{

st \%>\%
  stack_eval(mtcars) \%>\%
  stack_rm(spline_res_)
}

# to alter an evaluated model stack, unevaluate the
# model before adding or removing models
st_eval <- st \%>\%
  stack_eval(mtcars)
  
st_eval \%>\%
  stack_uneval() \%>\%
  stack_rm(spline_res_)

}
