% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fit.R
\name{stack_fit}
\alias{stack_fit}
\title{Fit model stack members}
\usage{
stack_fit(model_stack, data = NULL, ...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{model_stack} object with a subclass inherited from the chosen
\verb{*_stack} method---this fitted model contains the
necessary components to predict on new data.
}
\description{
Fit model stack members
}
\section{Example Data}{


See \code{?example_data} to learn more about the \code{lin_reg_res},
\code{reg_res_svm}, and \code{reg_res_sp} objects used in the examples below.
Long story short, they're resample objects outputted from
\code{\link[tune:fit_resamples]{tune::fit_resamples()}} or \code{\link[tune:tune_grid]{tune::tune_grid()}} fitting \code{body_mass_g}
with all other variables in \link[palmerpenguins:penguins]{palmerpenguins::penguins} as predictors!
}

\examples{

# put together a data stack
st <- 
  stacks() \%>\%
  stack_add(reg_res_lr) \%>\%
  stack_add(reg_res_svm) \%>\%
  stack_add(reg_res_sp)

# evaluate the data stack and fit the member models
st \%>\%
  stack_blend() \%>\%
  stack_fit()

}
