% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_predictions.R
\name{stack_predictions}
\alias{stack_predictions}
\title{Collate predictions from resampling results}
\usage{
stack_predictions(data, ..., metric = "rmse", n = 1)
}
\arguments{
\item{data}{Training data (as a \code{data.frame} or \code{tibble})}

\item{...}{Named resampling results (set of sub-models) outputted from
\code{tune_grid()}, where the name is an arbitrary and unique label for the
given set of sub-models. If \code{n > 1}, an integer will be appended to each
element sub-model in the set.}

\item{metric}{A character value giving the metric that will be used to sort
the models. (See
\url{https://tidymodels.github.io/yardstick/articles/metric-types.html} for
more details).}

\item{n}{An integer for the number of top sub-models to return. Either of
length 1 if the same number of sub-models should be included from each
set or of length equal to the number of sub-model set to specify how many
sub-models to include from each sub-model set.}
}
\value{
A tibble with \code{nrow(data)} rows and \verb{1 + <number of sub-models>}
columns, where each column (besides the first, which contains the true
response) gives the prediction for each selected sub-model.
}
\description{
Collate predictions from resampling results
}
\examples{
# just one sub-model from one sub-model set
stack_predictions(ames_train, glmnet = glmnet_res)

# to include results from 5 sub-models from the glmnet sub-model set and 1 
# from mars: 
stack_predictions(
  ames_train, 
  glmnet = glmnet_res,
  mars = mars_res,
  n = c(5, 1)
)

}
