% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks}
\alias{stacks}
\title{Initialize a resample stack}
\usage{
stacks(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{resample_stack} object.
}
\description{
Initializes a \code{resample_stack} object.
}
\seealso{
\code{\link[=stack_resamples]{stack_resamples()}}
}
