% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks}
\alias{stacks}
\title{Initialize a Stack}
\usage{
stacks(...)
}
\arguments{
\item{...}{Additional arguments. Currently ignored.}
}
\value{
A \code{data_stack} object.
}
\description{
The \code{stacks()} function initializes a \code{data_stack} object. Principally,
\code{data_stack}s are tibbles, where the first column gives
the true outcome in the assessment set, and the remaining
columns give the predictions from each candidate ensemble
member. (When the outcome is numeric, there’s only one column per candidate
member. For classification, there are as many columns per candidate
member as there are levels in the outcome variable minus 1.) They also bring
along a few extra attributes to keep track of model definitions, resamples,
and training data.

See \code{?stacks_description} for more discussion of the package, generally,
and the \code{basics} vignette for a detailed walk-through of functionality.
}
\seealso{
Other core verbs: 
\code{\link{add_candidates}()},
\code{\link{blend_predictions}()},
\code{\link{fit_members}()}
}
\concept{core verbs}
